moved parsing logic into schema, updated property interface, fixed typo in readme, scoped npm module name to @bonito
