+++ src/schema.ts
@@ -3,7 +3,6 @@ import dotty = require('dotty');
 
 import Property, { PropertyDefinition } from './property';
 import Type, { TypeDefinition } from './type';
-import parse from './parser';
 
 import numberType from './types/number';
 import booleanType from './types/boolean';
@@ -17,15 +16,47 @@ export interface HydrateOptionsDefinition {
   validate?: boolean;
 }
 
+export interface SchemaDefinition {
+  [optName: string]: any;
+}
+
+export interface ParsedSchemaDefinition {
+  [optName: string]: Property;
+}
+
 export default class Schema {
-  definition: Object;
+  definition: ParsedSchemaDefinition;
+
+  static parse(rawDefinition: SchemaDefinition, startingPath?: String): ParsedSchemaDefinition {
+    // 1. Loop through keys on object
+    // 2. Determine if value assigned to key is a Property or a nested object
+    // 3. Convert properties to Property objects
+    // 4. Parse nested values recursively
+
+    const parsedObj: ParsedSchemaDefinition = {};
+
+    Object.keys(rawDefinition).forEach((key) => {
+      const definition = rawDefinition[key];
+      const fullPath = startingPath ? [startingPath, key].join('.') : key;
+
+      if (Property.isProperty(definition)) {
+        parsedObj[fullPath] = new Property(fullPath, definition);
+      } else if (Property.isNested(definition)) {
+        Object.assign(parsedObj, Schema.parse(definition, fullPath));
+      } else {
+        throw new Error(`Property definition for ${fullPath} is illegal`);
+      }
+    });
+
+    return parsedObj;
+  }
 
   static addType(name: string, definition: TypeDefinition): void {
     Type.set(name, definition);
   }
 
-  constructor(rawDefinition: Object) {
-    this.definition = parse(rawDefinition);
+  constructor(rawDefinition: SchemaDefinition) {
+    this.definition = Schema.parse(rawDefinition);
   }
 
   hydrate(rawConfig: Object, options: HydrateOptionsDefinition = {}): Object {
